services/math-engine/requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
bcrypt==4.0.1
certifi==2026.2.25
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fakeredis==2.34.1
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
limits==5.8.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.9.2
mpmath==1.3.0
numpy>=1.26.0
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.20
redis==7.3.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
starlette==0.52.1
stripe==12.1.0
sympy==1.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
wrapt==2.1.2
psycopg2-binary==2.9.11
